LayoutTests/ipc/restrictedendpoints/allow-access-webGPU-expected.txt

CONSOLE MESSAGE: PASS: Test runner did not detect GPUP crash.
